ISSUE #1847 - Fix backend circ dep
