+++ backend/models/riskMitigation.js
@@ -17,6 +17,7 @@
 "use strict";
 
 const _ = require("lodash");
+const parse = require("csv-parse/lib/sync");
 const db = require("../handler/db");
 const responseCodes = require("../response_codes.js");
 const TeamspaceSettings = require("./teamspaceSetting");
@@ -92,6 +93,34 @@ class RiskMitigation {
 		const suggestions = await mitigationColl.find(criteria).toArray();
 
 		return suggestions;
+	}
+
+	async importCSV(account, data) {
+		const csvFields = [
+			"mitigation_desc",
+			"mitigation_detail",
+			"mitigation_stage",
+			"mitigation_type",
+			"category",
+			"location_desc",
+			"element",
+			"risk_factor",
+			"scope",
+			"associated_activity"
+		];
+
+		// remove column headers defined in template
+		data = data.replace(/Treatment Title,Treatment Details,Treatment Stage,Treatment Type,Risk Category,Risk Location,Element Type,Risk Factor,Construction Scope,Associated Activity\r\n/gm,"");
+
+		const records = parse(data, {
+			columns: csvFields,
+			skip_empty_lines: true,
+			trim: true
+		});
+
+		await this.clearAll(account);
+
+		return this.insert(account, records);
 	}
 
 	async insert(account, mitigations) {
